download: allow apikey via url path this is STRONGLY discouraged
Implement specifying apiKey in the url path. This allows legacy clients that are
unable to specify the apiKey via a header to retrieve keys. Due to the
potential security risks, this behavior is disabled by default.
Future commits will provide a construct to enable or disable this via
a setting saved in the db.