acme: add more robust checking to cert dl and also ability to specify preferred chain This does not make the preferred chain function available to end users, yet, but updates the acme package to support this ability.
The field to elect a preferred chain will behave the same way as certbot's currently does. If the preferred chain is found, that chain will be downloaded and saved. If it is not found, a default chain will instead be downloaded.

This also adds more robust cert chain checking. If the acme server returns a cert chain that is bad or incompatible with this client, it will now also check alt chains to try and recover by using a valid one of those.