private_keys: fix key pem standardization
In rare cases, an extra LF would be added to the end of the private key pem content causing an extra blank line in the pem content.

This occurred when the final line of the pem content was exactly 64 runes long.

This change prevents the extra LF from being added.

fixes: https://github.com/gregtwallace/certwarden/issues/56