orders: update auto ordering logic * remove 'valid_remaining_days_threshold' config option * instead calculate time to renew based on the percentage of validity remaining * include a backstop validity minimum value (e.g. if the backstop is hit before the percentage, do renewal) * change order ValidFrom and ValidTo to time.Time instead of int (all time should have this done eventually, but doing it here since I am already modifying this logic) * change db query when retrieving certs for renewal. to avoid logic in storage, pull all current valid orders and then filter them in the Orders package instead * remove unneeded db query * add storage helper func to convert db ints into time.Time
Goal here is to make auto-renewal play nice regardless of the validity period an ACME Server uses. If a Server is very strict and only issues very short certificates, renew when validity hits 10 days remaining regardless of percentage. Therefore, it is possible if a server issues something like a 7 day cert, the renewal will occur daily.

The percentage of validity and minimum backstop values are hardcoded and can be adjusted in the future if need be.