storage: fix v9 data migration edge case
Fixes an edge case where if the subject is a wildcard, it should not be copied to the client address (since a wildcard would be an invalid address).
Generally, this would not trigger since it is unlikely a user would have a client key generated where the subject is an invalid client address.

This does not cause divergence in the user population and therefore can be applied in the existing v9 function, as opposed to doing an additional migration.